fix: disable plugin execution when '-*' is used as exception
When there is an exception targeting a whole plugin (eg
`pycodestyle = ['-*']`) and the plugin is activated globally (eg
`pycodestyle = ['+*']`), the plugin would be executed anyway.

This is  because the `_should_process` method checks for
`not rules or set(rules) == {'-*'}`.

This change replaces the check with the following procedure:

If `'-*'` is a valid exception, the plugin configuration equals
the respective exception (minus the `'-*'` glob, of course).
Otherwise, the exception config is appended to the plugin config
list.

An implication of this is that if an exception contains only `-*`,
and it matches the current file, `_should_process` returns `False`
and the plugin check is skipped from execution entirely.

fixes #149